requirements.txt
django
gunicorn==19.7.1
django-heroku
requests
psycopg2-binary
pandas
argon2_cffi
